pass env to deployment

+++ deploy.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 set -e
+if [ ! "${SQS_URL+set}"]; then
+  echo "set env SQS_URL"
+  exit 1
+fi
+
+docker-compose -f docker-compose.yml -f docker-compose.prod.yml ps
 docker-compose -f docker-compose.yml -f docker-compose.prod.yml build
 docker-compose -f docker-compose.yml -f docker-compose.prod.yml down
-docker-compose -f docker-compose.yml -f docker-compose.prod.yml up -d
+docker-compose -f docker-compose.yml -f docker-compose.prod.yml up -d -e SQS_URL=$SQS_URL
+docker-compose -f docker-compose.yml -f docker-compose.prod.yml ps
